split mode: Hack the layout to be an even 50/50 split
Also does a hack to make the layout width for the divider to be 80px wide, which
only works on a Surface Duo first generation device.

What we really need to do is make K-9 multi-process and have two different
activites running, one on each screen instead of hacking the split view for
this.