message list options: Rearrange actions in order of usage
This aligns the icons in the order of the "Visible message actions" config menu,
while also always making the most frequently used options visible action bar
when there's space by using "always". Android's Toolbar class has a tendency to
limit the number of visible actions to 4, even when there is space. Instead, we
have to mark these actions as being "always" visible outside of the extra menu
to force it to grow beyond that limit.

As a side-effect of this, MessageList options are now appended to the end of the
action list. So in split view mode, message list actions are kept in the
overflow menu instead of spilling out to the MessageItem toolbar, so that
actions that are more contextually relevant are visible.